% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.webmd.obgyn-package.R
\docType{package}
\name{dc.webmd.obgyn-package}
\alias{dc.webmd.obgyn}
\alias{dc.webmd.obgyn-package}
\title{dc.webmd.obgyn: Social Data Commons: Obstetrics and Gynecologists Access Scores}
\description{
Allows user to easily get YOUR_DATA data from the SDAD Data Commons.  Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Cesar Montalvo \email{cpm9w@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
